Add sharding rule visitor
